LatticePointNeighbourhoodPosition: Factory method pattern
Object Method: getNeighbourhoodFor
description:The method is a relationship that relies directly on a switch according to the laticeNeighbourhood type, so it encapsulates object generation using factory method patterns to increase code flexibility and prevent client code from relying on specific classes. When you apply a factory method pattern, passing a neighbor type creates an array of neighbor locations for that type and returns it.