LatticePoint: Extract Array of Offsets
Object method: getNeighbourhood

Description:  It involves extracting the repeated offsets into a separate array structure, making the code more concise and easier to manage. Instead of manually assigning the offsets for each neighbour point, the offsets are stored in a 2D array (offsets) where each row represents the x and y offsets for a specific neighbour point. By iterating over this array, we can calculate the new coordinates for each neighbour point in a more systematic and maintainable way