CellCoreOriginal.java : Factory Method Pattern
It was originally CellCore.java.

As Factory Method Pattern applied, it renamed to CellCoreOriginal(to
indicate it is the original logic of CellCore).

Factory Method Pattern was applied, so it's fine to add new CellCore class such
as CellCoreEvolution, CellCoreDegeneration which has different CellCore
logic.

And It satisfies OCP when it comes to add such new CellCore classes.